feat: mv project stakeholders to details section
